Commit #14
Now questions are being properly displayed using the viewModel + LiveData + encapsulation

Next step is to implement the logic for detecting right answers and deciding what to do
